+++ tests/testthat/test-prerun.R
@@ -12,8 +12,8 @@ test_that("import failures", {
                "Path 'spectrum.zip' for 'spectrum' does not exist")
   file.create(file.path(p, "spectrum.zip"))
   expect_error(obj$import(p),
-               "Path 'summary.zip' for 'summary' does not exist")
-  file.create(file.path(p, "summary.zip"))
+               "Path 'coarse-output.zip' for 'coarse_output' does not exist")
+  file.create(file.path(p, "coarse-output.zip"))
   expect_error(obj$import(p, "path/to/output.zip"),
                "Path for 'output' must be just the filename, no slashes")
 })
@@ -25,24 +25,25 @@ test_that("import base data", {
 
   p <- system_file("output")
 
-  inputs <- read_info_inputs(file.path(p, "malawi_summary_download.zip"))
+  inputs <- read_info_inputs(file.path(p, "malawi_coarse_output_download.zip"))
   expect_false(obj$exists(inputs))
 
   h <- obj$import(p, "malawi_output.rds",
                   "malawi_spectrum_download.zip",
-                  "malawi_summary_download.zip")
+                  "malawi_coarse_output_download.zip")
   expect_equal(obj$list(), h)
   expect_true(obj$exists(inputs))
   expect_equal(obj$get(inputs),
                list(output_path = file.path(path_prerun, h, "output.rds"),
                     spectrum_path = file.path(path_prerun, h, "spectrum.zip"),
-                    summary_path = file.path(path_prerun, h, "summary.zip")))
+                    coarse_output_path = file.path(path_prerun, h,
+                                                   "coarse-output.zip")))
   expect_true(all(vapply(obj$get(inputs), file.exists, TRUE)))
 
   expect_error(
     obj$import(p, "malawi_output.rds",
                "malawi_spectrum_download.zip",
-               "malawi_summary_download.zip"),
+               "malawi_coarse_output_download.zip"),
     "This set of data has been imported already")
 })
 
@@ -54,7 +55,7 @@ test_that("run with prerun", {
   h <- prerun_import(path_prerun, system_file("output"),
                      "malawi_output.rds",
                      "malawi_spectrum_download.zip",
-                     "malawi_summary_download.zip")
+                     "malawi_coarse_output_download.zip")
 
   data <- list(
     pjnz = list(filename = "Malawi2019.PJNZ",
@@ -125,7 +126,7 @@ test_that("run with prerun", {
   args <- c(path_prerun, system_file("output"),
             "--output=malawi_output.rds",
             "--spectrum=malawi_spectrum_download.zip",
-            "--summary=malawi_summary_download.zip")
+            "--coarse-output=malawi_coarse_output_download.zip")
   expect_message(main_import_prerun(args),
                  "Imported data as '[[:xdigit:]]+'")
   expect_equal(length(obj$list()), 1)